backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.45.1
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.30
requests==2.32.3
joblib==1.4.2
python-dotenv==1.0.1
pytest==8.2.0
